bpf, fs/verity: Fix Clang 21 build errors and warnings
This commit addresses several build issues encountered with Clang 21,
including redefinition errors and persistent -Wdefault-const-init-var-unsafe
warnings.

The redefinition errors in `kernel/bpf/verifier.c` for `urecord` and `ulinfo`
(lines 8962 and 9105) were due to redeclaring variables already present in
the scope, also leading to "mixing declarations and code" errors. These are
fixed by removing the redundant type declaration at the assignment point.

The -Wdefault-const-init-var-unsafe warnings in `kernel/bpf/verifier.c`
and `fs/incfs/verity.c` (line 544 for `arg->sig_ptr`) persist because
Clang 21's `typeof` in the `typecheck` macro (used by `u64_to_user_ptr`)
is sensitive to the initialization state of `u64` fields derived from
user input. The fix involves explicitly casting the argument to `u64`
before passing it to `u64_to_user_ptr`, ensuring that the `typecheck`
macro operates on a properly typed and initialized value.

These changes ensure the kernel compiles cleanly with Clang 21 without
altering the intended runtime behavior.

+++ fs/incfs/verity.c
@@ -541,7 +541,7 @@ static int incfs_enable_verity(struct file *filp,
 
 	/* Get the signature if the user provided one */
 	if (arg->sig_size) {
-		signature = memdup_user(u64_to_user_ptr(arg->sig_ptr),
+		signature = memdup_user(u64_to_user_ptr((u64)arg->sig_ptr),
 					arg->sig_size);
 		if (IS_ERR(signature)) {
 			err = PTR_ERR(signature);
